+++ crates/lynx-db/src/dao/request_processing_dao/handlers/mod.rs
@@ -1,5 +1,6 @@
 pub mod block_handler;
 pub mod handler_rule;
+pub mod html_script_injector;
 pub mod local_file_handler;
 pub mod modify_request_handler;
 pub mod modify_response_handler;
@@ -7,5 +8,6 @@ pub mod proxy_forward_handler;
 
 pub use block_handler::BlockHandlerConfig;
 pub use handler_rule::HandlerRule;
+pub use html_script_injector::HtmlScriptInjectorConfig;
 pub use local_file_handler::LocalFileConfig;
 pub use modify_request_handler::ModifyRequestConfig;