+++ crates/lynx-mock/examples/server.rs
@@ -8,7 +8,7 @@ async fn main() -> Result<()> {
         .with(fmt::layer())
         .with(EnvFilter::from_default_env().add_directive("lynx_mock=trace".parse()?))
         .init();
-    let mut server = MockServer::new(Some(3000));
+    let mut server = MockServer::new(Some(7788));
     server.write_cert_to_file()?;
     server.start_server().await?;
     let client = MockClient::new(Some(vec![server.cert.clone()]), None)?;