+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.2.4] - 2025-06-13
+## [unreleased]
 
 ### 🚀 Features
 
@@ -10,6 +10,8 @@ All notable changes to this project will be documented in this file.
 - The command line supports daemons ([#15](https://github.com/suxin2017/lynx-server/issues/15))
 - Pwa and one click add block rule ([#23](https://github.com/suxin2017/lynx-server/issues/23))
 - Support jaeger log ([#25](https://github.com/suxin2017/lynx-server/issues/25))
+- Better display of status time ([#28](https://github.com/suxin2017/lynx-server/issues/28))
+- Support inject script ([#29](https://github.com/suxin2017/lynx-server/issues/29))
 
 ### 🐛 Bug Fixes